float you punk
git-svn-id: svn+ssh://nlp.cs.jhu.edu/home/remote/asheesh/collab@259 fe7ecb3d-170a-0410-af80-b42c727aa7d5

+++ code/lib/style_edit_finding.py
@@ -100,7 +100,7 @@ def plus_and_minus2hunks(s):
         # Look for a number line first.  
         if line[0] != '-' and line[0] != '+' and plusses: # then reset state
             yield HunkOfSentences(olds=minuses, news=plusses, confidence=confid)
-            confid = int(line)
+            confid = float(line)
             minuses, plusses = [], []
         elif line[0] == '-':
             minuses.append(line[1:])
